Move HTTP REST to api.py for unified connection port.
